Merge branch 'Work_Some_Broke_Some' of https://github.com/xrofng/Fisherman-Friend into Animation

# Conflicts:
#	Fish Frenzy/Assets/O-Asset/LowPolyWater_Pack/Plane Meshes/Ocean50x50W750H750.asset

+++ Fish Frenzy/Assets/Script/Player/PlayerFishSpecial.cs
@@ -74,7 +74,7 @@ public class PlayerFishSpecial : PlayerAbility
                   () => { specialTrail.gameObject.SetActive(true); specialTrail.Play(); },
                   () => { specialTrail.gameObject.SetActive(false); specialTrail.Stop(); });
 
-            _pAnimator.ChangeAnimState(PlayerAnimation.State.V_Slap, FishSpecial<FishSpecialMelee>().SpeiclaClipFrameCount, true, PlayerAnimation.State.HoldFish);             
+            _pAnimator.ChangeAnimState((int)PlayerAnimation.State.V_Slap, FishSpecial<FishSpecialMelee>().SpeiclaClipFrameCount, true, (int)PlayerAnimation.State.HoldFish);             
         }   
     }
